requirements.txt
Flask==3.0.1
Flask-JWT-Extended==4.6.0
validate_email_address==1.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.3 
pgvector==0.2.4
langchain-community==0.0.1
langchain==0.1.4
tiktoken==0.5.2
langchain-openai==0.0.5